// src/types/schemas/node--page.d.ts
export type ID = number;
export type RevisionID = number;
export type Language = string;
/**
 * The time that the current revision was created.
 */
export type RevisionCreateTime = number;
/**
 * Briefly describe the changes you have made.
 */
export type RevisionLogMessage = string;
export type Published = boolean;
export type Title = string;
/**
 * The time that the node was created.
 */
export type AuthoredOn = number;
/**
 * The time that the node was last edited.
 */
export type Changed = number;
export type PromotedToFrontPage = boolean;
export type StickyAtTopOfLists = boolean;
/**
 * A flag indicating whether this is the default translation.
 */
export type DefaultTranslation = boolean;
/**
 * A flag indicating whether this was a default revision when it was saved.
 */
export type DefaultRevision = boolean;
/**
 * Indicates if the last edit of a translation belongs to current revision.
 */
export type RevisionTranslationAffected = boolean;
/**
 * The moderation state of this piece of content.
 */
export type ModerationState = string;
export type PathAlias = string;
export type PathId = number;
export type LanguageCode = string;
export type Text = string;
export type TextFormat = string;
export type Summary = string;
// ...
export interface Attributes {
  drupal_internal__nid?: ID;
  drupal_internal__vid?: RevisionID;
  langcode?: Language;
  revision_timestamp?: RevisionCreateTime;
  revision_log?: RevisionLogMessage;
  status?: Published;
  title: Title;
  created?: AuthoredOn;
  changed?: Changed;
  promote?: PromotedToFrontPage;
  sticky?: StickyAtTopOfLists;
  default_langcode?: DefaultTranslation;
  revision_default?: DefaultRevision;
  revision_translation_affected?: RevisionTranslationAffected;
  moderation_state?: ModerationState;
  metatag?: MetatagsHiddenFieldForJSONSupport;
  path?: URLAlias;
  body?: Body;
}
/**
 * The meta tags for the entity.
 */
export interface MetatagsHiddenFieldForJSONSupport {
  [k: string]: unknown;
}
export interface URLAlias {
  alias?: PathAlias;
  pid?: PathId;
  langcode?: LanguageCode;
  [k: string]: unknown;
}
export interface Body {
  value: Text;
  format?: TextFormat;
  summary?: Summary;
  [k: string]: unknown;
}